+++ src/basic/components/CartPage.tsx
@@ -1,34 +1,14 @@
-// TODO: 장바구니 페이지 컴포넌트
-// 힌트:
-// 1. 상품 목록 표시 (검색 기능 포함)
-// 2. 장바구니 관리
-// 3. 쿠폰 적용
-// 4. 주문 처리
-//
-// 필요한 hooks:
-// - useProducts: 상품 목록 관리
-// - useCart: 장바구니 상태 관리
-// - useCoupons: 쿠폰 목록 관리
-// - useDebounce: 검색어 디바운싱
-//
-// 하위 컴포넌트:
-// - SearchBar: 검색 입력
-// - ProductList: 상품 목록 표시
-// - Cart: 장바구니 표시 및 결제
-
-import { useCallback, useEffect, useState } from 'react';
+import { useCallback, useState } from 'react';
 
 import type { CartItem, Coupon, NotificationVariant } from '../../types';
 import type { ProductWithUI } from '../constants';
 import { Icon } from './icons';
-import {
-  calculateCartTotal,
-  calculateItemTotal,
-  calculateItemTotalQuantity,
-  filterProducts,
-  getRemainingStock,
-} from '../models/entity';
-import { formatKRWPrice } from '../utils/formatters';
+import { calculateCartTotal, getRemainingStock } from '../models/entity';
+import { CartHeader } from './ui/CartHeader';
+import { CartItemList } from './ui/CartItemList';
+import { CouponSelector } from './ui/CouponSelector';
+import { PaymentSummary } from './ui/PaymentSummary';
+import { ProductList } from './ui/ProductList';
 import { useDebounce } from '../utils/hooks/useDebounce';
 import { useLocalStorage } from '../utils/hooks/useLocalStorage';
 
@@ -59,16 +39,6 @@ export function CartPage({
   const debouncedSearchTerm = useDebounce(searchTerm, 500);
 
   const [cart, setCart] = useLocalStorage<CartItem[]>('cart', []);
-  const [totalItemCount, setTotalItemCount] = useState(0);
-
-  useEffect(() => {
-    const count = calculateItemTotalQuantity(cart);
-    setTotalItemCount(count);
-  }, [cart]);
-
-  const filteredProducts = debouncedSearchTerm
-    ? filterProducts(products, debouncedSearchTerm)
-    : products;
 
   const applyCoupon = useCallback(
     (coupon: Coupon) => {
@@ -157,41 +127,12 @@ export function CartPage({
 
   return (
     <>
-      <header className='bg-white shadow-sm sticky top-0 z-40 border-b'>
-        <div className='max-w-7xl mx-auto px-4'>
-          <div className='flex justify-between items-center h-16'>
-            <div className='flex items-center flex-1'>
-              <h1 className='text-xl font-semibold text-gray-800'>SHOP</h1>
-              <div className='ml-8 flex-1 max-w-md'>
-                <input
-                  type='text'
-                  value={searchTerm}
-                  onChange={(e) => setSearchTerm(e.target.value)}
-                  placeholder='상품 검색...'
-                  className='w-full px-4 py-2 border border-gray-300 rounded-lg focus:outline-none focus:border-blue-500'
-                />
-              </div>
-            </div>
-            <nav className='flex items-center space-x-4'>
-              <button
-                onClick={() => setIsAdmin(true)}
-                className='px-3 py-1.5 text-sm rounded transition-colors text-gray-600 hover:text-gray-900'
-              >
-                관리자 페이지로
-              </button>
-              <div className='relative'>
-                <Icon name='cart' width={24} height={24} />
-
-                {cart.length > 0 && (
-                  <span className='absolute -top-2 -right-2 bg-red-500 text-white text-xs rounded-full w-5 h-5 flex items-center justify-center'>
-                    {totalItemCount}
-                  </span>
-                )}
-              </div>
-            </nav>
-          </div>
-        </div>
-      </header>
+      <CartHeader
+        searchTerm={searchTerm}
+        setSearchTerm={setSearchTerm}
+        setIsAdmin={setIsAdmin}
+        cart={cart}
+      />
 
       <main className='max-w-7xl mx-auto px-4 py-8'>
         <div className='grid grid-cols-1 lg:grid-cols-4 gap-6'>
@@ -202,93 +143,12 @@ export function CartPage({
                 <h2 className='text-2xl font-semibold text-gray-800'>전체 상품</h2>
                 <div className='text-sm text-gray-600'>총 {products.length}개 상품</div>
               </div>
-              {filteredProducts.length === 0 ? (
-                <div className='text-center py-12'>
-                  <p className='text-gray-500'>
-                    "{debouncedSearchTerm}"에 대한 검색 결과가 없습니다.
-                  </p>
-                </div>
-              ) : (
-                <div className='grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-4'>
-                  {filteredProducts.map((product) => {
-                    const remainingStock = getRemainingStock(product, cart);
-
-                    return (
-                      <div
-                        key={product.id}
-                        className='bg-white rounded-lg border border-gray-200 overflow-hidden hover:shadow-lg transition-shadow'
-                      >
-                        {/* 상품 이미지 영역 (placeholder) */}
-                        <div className='relative'>
-                          <div className='aspect-square bg-gray-100 flex items-center justify-center'>
-                            <Icon name='noImage' width={96} height={96} />
-                          </div>
-                          {product.isRecommended && (
-                            <span className='absolute top-2 right-2 bg-red-500 text-white text-xs px-2 py-1 rounded'>
-                              BEST
-                            </span>
-                          )}
-                          {product.discounts.length > 0 && (
-                            <span className='absolute top-2 left-2 bg-orange-500 text-white text-xs px-2 py-1 rounded'>
-                              ~{Math.max(...product.discounts.map((d) => d.rate)) * 100}%
-                            </span>
-                          )}
-                        </div>
-
-                        {/* 상품 정보 */}
-                        <div className='p-4'>
-                          <h3 className='font-medium text-gray-900 mb-1'>{product.name}</h3>
-                          {product.description && (
-                            <p className='text-sm text-gray-500 mb-2 line-clamp-2'>
-                              {product.description}
-                            </p>
-                          )}
-
-                          {/* 가격 정보 */}
-                          <div className='mb-3'>
-                            <p className='text-lg font-bold text-gray-900'>
-                              {remainingStock <= 0
-                                ? 'SOLD OUT'
-                                : formatKRWPrice(product.price, 'symbol')}
-                            </p>
-                            {product.discounts.length > 0 && (
-                              <p className='text-xs text-gray-500'>
-                                {product.discounts[0].quantity}개 이상 구매시 할인{' '}
-                                {product.discounts[0].rate * 100}%
-                              </p>
-                            )}
-                          </div>
-
-                          {/* 재고 상태 */}
-                          <div className='mb-3'>
-                            {remainingStock <= 5 && remainingStock > 0 && (
-                              <p className='text-xs text-red-600 font-medium'>
-                                품절임박! {remainingStock}개 남음
-                              </p>
-                            )}
-                            {remainingStock > 5 && (
-                              <p className='text-xs text-gray-500'>재고 {remainingStock}개</p>
-                            )}
-                          </div>
-
-                          {/* 장바구니 버튼 */}
-                          <button
-                            onClick={() => addToCart(product)}
-                            disabled={remainingStock <= 0}
-                            className={`w-full py-2 px-4 rounded-md font-medium transition-colors ${
-                              remainingStock <= 0
-                                ? 'bg-gray-100 text-gray-400 cursor-not-allowed'
-                                : 'bg-gray-900 text-white hover:bg-gray-800'
-                            }`}
-                          >
-                            {remainingStock <= 0 ? '품절' : '장바구니 담기'}
-                          </button>
-                        </div>
-                      </div>
-                    );
-                  })}
-                </div>
-              )}
+              <ProductList
+                products={products}
+                debouncedSearchTerm={debouncedSearchTerm}
+                cart={cart}
+                addToCart={addToCart}
+              />
             </section>
           </div>
 
@@ -299,146 +159,24 @@ export function CartPage({
                   <Icon name='cartBagRegular' width={20} height={20} className='mr-2' />
                   장바구니
                 </h2>
-                {cart.length === 0 ? (
-                  <div className='text-center py-8'>
-                    <Icon
-                      name='cartBagThin'
-                      width={64}
-                      height={64}
-                      className='text-gray-300 mx-auto mb-4'
-                    />
-                    <p className='text-gray-500 text-sm'>장바구니가 비어있습니다</p>
-                  </div>
-                ) : (
-                  <div className='space-y-3'>
-                    {cart.map((item) => {
-                      const itemTotal = calculateItemTotal(item, cart);
-                      const originalPrice = item.product.price * item.quantity;
-                      const hasDiscount = itemTotal < originalPrice;
-                      const discountRate = hasDiscount
-                        ? Math.round((1 - itemTotal / originalPrice) * 100)
-                        : 0;
 
-                      return (
-                        <div key={item.product.id} className='border-b pb-3 last:border-b-0'>
-                          <div className='flex justify-between items-start mb-2'>
-                            <h4 className='text-sm font-medium text-gray-900 flex-1'>
-                              {item.product.name}
-                            </h4>
-                            <button
-                              onClick={() => removeFromCart(item.product.id)}
-                              className='text-gray-400 hover:text-red-500 ml-2'
-                            >
-                              <Icon name='x' width={16} height={16} />
-                            </button>
-                          </div>
-                          <div className='flex items-center justify-between'>
-                            <div className='flex items-center'>
-                              <button
-                                onClick={() => updateQuantity(item.product.id, item.quantity - 1)}
-                                className='w-6 h-6 rounded border border-gray-300 flex items-center justify-center hover:bg-gray-100'
-                              >
-                                <span className='text-xs'>−</span>
-                              </button>
-                              <span className='mx-3 text-sm font-medium w-8 text-center'>
-                                {item.quantity}
-                              </span>
-                              <button
-                                onClick={() => updateQuantity(item.product.id, item.quantity + 1)}
-                                className='w-6 h-6 rounded border border-gray-300 flex items-center justify-center hover:bg-gray-100'
-                              >
-                                <span className='text-xs'>+</span>
-                              </button>
-                            </div>
-                            <div className='text-right'>
-                              {hasDiscount && (
-                                <span className='text-xs text-red-500 font-medium block'>
-                                  -{discountRate}%
-                                </span>
-                              )}
-                              <p className='text-sm font-medium text-gray-900'>
-                                {Math.round(itemTotal).toLocaleString()}원
-                              </p>
-                            </div>
-                          </div>
-                        </div>
-                      );
-                    })}
-                  </div>
-                )}
+                <CartItemList
+                  cart={cart}
+                  onRemoveFromCart={removeFromCart}
+                  onUpdateQuantity={updateQuantity}
+                />
               </section>
 
               {cart.length > 0 && (
                 <>
-                  <section className='bg-white rounded-lg border border-gray-200 p-4'>
-                    <div className='flex items-center justify-between mb-3'>
-                      <h3 className='text-sm font-semibold text-gray-700'>쿠폰 할인</h3>
-                      <button className='text-xs text-blue-600 hover:underline'>쿠폰 등록</button>
-                    </div>
-                    {coupons.length > 0 && (
-                      <select
-                        className='w-full text-sm border border-gray-300 rounded px-3 py-2 focus:outline-none focus:border-blue-500'
-                        value={selectedCoupon?.code || ''}
-                        onChange={(e) => {
-                          const coupon = coupons.find((c) => c.code === e.target.value);
-                          if (coupon) applyCoupon(coupon);
-                          else setSelectedCoupon(null);
-                        }}
-                      >
-                        <option value=''>쿠폰 선택</option>
-                        {coupons.map((coupon) => (
-                          <option key={coupon.code} value={coupon.code}>
-                            {coupon.name} (
-                            {coupon.discountType === 'amount'
-                              ? `${coupon.discountValue.toLocaleString()}원`
-                              : `${coupon.discountValue}%`}
-                            )
-                          </option>
-                        ))}
-                      </select>
-                    )}
-                  </section>
+                  <CouponSelector
+                    coupons={coupons}
+                    selectedCoupon={selectedCoupon}
+                    onApplyCoupon={applyCoupon}
+                    onResetSelectedCoupon={() => setSelectedCoupon(null)}
+                  />
 
-                  <section className='bg-white rounded-lg border border-gray-200 p-4'>
-                    <h3 className='text-lg font-semibold mb-4'>결제 정보</h3>
-                    <div className='space-y-2 text-sm'>
-                      <div className='flex justify-between'>
-                        <span className='text-gray-600'>상품 금액</span>
-                        <span className='font-medium'>
-                          {totals.totalBeforeDiscount.toLocaleString()}원
-                        </span>
-                      </div>
-                      {totals.totalBeforeDiscount - totals.totalAfterDiscount > 0 && (
-                        <div className='flex justify-between text-red-500'>
-                          <span>할인 금액</span>
-                          <span>
-                            -
-                            {(
-                              totals.totalBeforeDiscount - totals.totalAfterDiscount
-                            ).toLocaleString()}
-                            원
-                          </span>
-                        </div>
-                      )}
-                      <div className='flex justify-between py-2 border-t border-gray-200'>
-                        <span className='font-semibold'>결제 예정 금액</span>
-                        <span className='font-bold text-lg text-gray-900'>
-                          {totals.totalAfterDiscount.toLocaleString()}원
-                        </span>
-                      </div>
-                    </div>
-
-                    <button
-                      onClick={completeOrder}
-                      className='w-full mt-4 py-3 bg-yellow-400 text-gray-900 rounded-md font-medium hover:bg-yellow-500 transition-colors'
-                    >
-                      {totals.totalAfterDiscount.toLocaleString()}원 결제하기
-                    </button>
-
-                    <div className='mt-3 text-xs text-gray-500 text-center'>
-                      <p>* 실제 결제는 이루어지지 않습니다</p>
-                    </div>
-                  </section>
+                  <PaymentSummary totals={totals} completeOrder={completeOrder} />
                 </>
               )}
             </div>
