+++ src/basic/components/ui/CartItemList.tsx
@@ -1,0 +1,70 @@
+import type { CartItem } from '../../../types';
+import { calculateItemTotal } from '../../models/entity';
+import { Icon } from '../icons';
+
+interface CartItemListProps {
+  cart: CartItem[];
+  onRemoveFromCart: (productId: string) => void;
+  onUpdateQuantity: (productId: string, quantity: number) => void;
+}
+
+export function CartItemList({ cart, onRemoveFromCart, onUpdateQuantity }: CartItemListProps) {
+  if (cart.length === 0) {
+    return (
+      <div className='text-center py-8'>
+        <Icon name='cartBagThin' width={64} height={64} className='text-gray-300 mx-auto mb-4' />
+        <p className='text-gray-500 text-sm'>장바구니가 비어있습니다</p>
+      </div>
+    );
+  }
+
+  return (
+    <div className='space-y-3'>
+      {cart.map((item) => {
+        const itemTotal = calculateItemTotal(item, cart);
+        const originalPrice = item.product.price * item.quantity;
+        const hasDiscount = itemTotal < originalPrice;
+        const discountRate = hasDiscount ? Math.round((1 - itemTotal / originalPrice) * 100) : 0;
+
+        return (
+          <div key={item.product.id} className='border-b pb-3 last:border-b-0'>
+            <div className='flex justify-between items-start mb-2'>
+              <h4 className='text-sm font-medium text-gray-900 flex-1'>{item.product.name}</h4>
+              <button
+                onClick={() => onRemoveFromCart(item.product.id)}
+                className='text-gray-400 hover:text-red-500 ml-2'
+              >
+                <Icon name='x' width={16} height={16} />
+              </button>
+            </div>
+            <div className='flex items-center justify-between'>
+              <div className='flex items-center'>
+                <button
+                  onClick={() => onUpdateQuantity(item.product.id, item.quantity - 1)}
+                  className='w-6 h-6 rounded border border-gray-300 flex items-center justify-center hover:bg-gray-100'
+                >
+                  <span className='text-xs'>−</span>
+                </button>
+                <span className='mx-3 text-sm font-medium w-8 text-center'>{item.quantity}</span>
+                <button
+                  onClick={() => onUpdateQuantity(item.product.id, item.quantity + 1)}
+                  className='w-6 h-6 rounded border border-gray-300 flex items-center justify-center hover:bg-gray-100'
+                >
+                  <span className='text-xs'>+</span>
+                </button>
+              </div>
+              <div className='text-right'>
+                {hasDiscount && (
+                  <span className='text-xs text-red-500 font-medium block'>-{discountRate}%</span>
+                )}
+                <p className='text-sm font-medium text-gray-900'>
+                  {Math.round(itemTotal).toLocaleString()}원
+                </p>
+              </div>
+            </div>
+          </div>
+        );
+      })}
+    </div>
+  );
+}
